+++ app.js
@@ -4,19 +4,13 @@ const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
 const cors = require('cors');
 const cookieParser = require('cookie-parser');
+const helmet = require('helmet');
 const { errors } = require('celebrate');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 const appRoutes = require('./routes');
+const { limiter, allowedCors } = require('./utils/constants/constants');
 
 const { DB_URL = 'mongodb://localhost:27017/moviesdb' } = process.env;
-const allowedCors = [
-  'http://localhost:3000',
-  'https://localhost:3000',
-  'http://felipsewindchaser.nomoredomains.sbs',
-  'https://felipsewindchaser.nomoredomains.sbs',
-  'http://api.felipsewindchaser.nomoredomains.sbs',
-  'https://api.felipsewindchaser.nomoredomains.sbs',
-];
 const { PORT = 3000 } = process.env;
 const app = express();
 
@@ -24,6 +18,7 @@ mongoose.connect(DB_URL, {
   useNewUrlParser: true,
   autoIndex: true,
 });
+app.use(helmet());
 app.use(requestLogger);
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -34,6 +29,7 @@ app.use(
   }),
 );
 app.use(cookieParser());
+app.use(limiter);
 app.get('/crash-test', () => {
   setTimeout(() => {
     throw new Error('Сервер сейчас упадёт');
