+++ package.json
@@ -17,7 +17,9 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
+    "express-rate-limit": "^6.5.1",
     "express-winston": "^4.2.0",
+    "helmet": "^5.1.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.5.0",
     "validator": "^13.7.0",
